requirements.txt
anyio==4.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.1
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==1.0.0
python-telegram-bot==20.7
requests==2.31.0
rsa==4.9
sniffio==1.3.0
tqdm==4.66.1
urllib3==2.1.0
Pillow==10.1.0